WIX 111
#Fix Ticket #111
-  Update AiModel Table.
- Updated askCompletion to take as Input aiModel. Changed the appropriate files.
- Made a general Gptrequest, response and message.
- Trying to add GptConfig that takes as inputs top_p, max_token, temperature.

### Related Ticket: #111

+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/services/ProjectAgentService.java
@@ -26,9 +26,9 @@ public class ProjectAgentService {
     @Value("${gendox.agents.max_token}")
     private Long maxTokenValue;
     @Value("${gendox.agents.temperature}")
-    private Float temperatureValue;
+    private Double temperatureValue;
     @Value("${gendox.agents.top_p}")
-    private Float topPValue;
+    private Double topPValue;
 
     private ProjectAgentRepository projectAgentRepository;
     private TypeService typeService;
